last minute changes
