+++ src/components/Footer.js
@@ -1,36 +1,8 @@
-import React, { useState } from "react"
-import Sound from "react-sound"
+import React from "react"
 
 const Footer = () => {
-  const [isPlaying, setIsPlaying] = useState(false)
-  const audioSrc = "/gentlewaves.mp3"
-
-  const togglePlayPause = () => {
-    const audio = document.getElementById("audio-player")
-    if (isPlaying) {
-      audio.pause()
-    } else {
-      audio.play()
-    }
-    setIsPlaying(!isPlaying)
-  }
   return (
     <footer className="footer-text">
-      <div className="audio-player-container">
-        <audio
-          id="audio-player"
-          src={audioSrc}
-          loop // This makes the audio loop
-          onPlay={() => setIsPlaying(true)}
-          onPause={() => setIsPlaying(false)}
-        >
-          Your browser does not support the audio element.
-        </audio>
-        <button onClick={togglePlayPause}>
-          {isPlaying ? "Pause" : "Play"}
-        </button>
-      </div>
-
       <p style={{ color: "#e7b47f" }}>
         © 2024{" "}
         <a
